requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.6.3
pyTelegramBotAPI==3.8.1
python-dotenv==0.18.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
